[hipdnn] Drop PowerShell wrapper and adopt log-on-failure pattern in skills
Addresses two review comments from Brian on the superbuild skills PR.

The Windows DLL search path can be set entirely via Python's subprocess
env block — CreateProcessW propagates it to grandchildren (cmake -> ninja
-> ctest -> test.exe), so the prior PowerShell wrapper isn't needed.
Verified end-to-end on Windows against miopen-provider, hipblaslt-provider,
and hipdnn-integration-tests unit-checks (no STATUS_DLL_NOT_FOUND).

Renames pwsh_cmake_run.py to cmake_run.py and rewrites it to:
- Set PATH and ROCM_PATH on the subprocess env block, no shell-out.
- Eliminate the f-string interpolation of paths/targets into a PowerShell
  command string, which broke on paths with spaces and was injection-prone
  for paths containing single quotes or PowerShell metacharacters.
- Add a --binary mode (with --gtest-filter) so Step 5b can use the same
  helper instead of hand-rolling another powershell -Command line.

Replaces the `... 2>&1 | tail -100` pattern in both SKILL.md files (build
step and test steps) with a log-and-tail-on-failure form that captures
full output to a temp file, tails only on failure, and propagates the
real exit code. Bash without pipefail returns tail's exit code from a
pipeline, which would silently mask test segfaults or DLL-load crashes
that produce no "FAIL" string for the orchestrator to parse.
